refactored lib
