+++ flow/pyproject.toml
@@ -19,13 +19,12 @@ pycryptodome = "^3.20.0"
 sqlalchemy = "^2.0.25"
 asyncpg = "^0.29.0"
 pydantic = "^2.5.3"
-azure-storage-blob = "^12.19.0"
 azure-identity = "^1.15.0"
 aiofiles = "^23.2.1"
-aiohttp = "^3.9.4"
 psycopg2-binary = "^2.9.9"
 
 [tool.poetry.group.dev.dependencies]
+lib = {path = "../lib", develop = true}
 fastapi = "^0.109.0"
 uvicorn = "^0.26.0"
