+++ indexer/pyproject.toml
@@ -23,6 +23,8 @@ fitz = "^0.0.1.dev2"
 aiofiles = "^23.2.1"
 confluent-kafka = "^2.3.0"
 
+[tool.poetry.group.dev.dependencies]
+lib = {path = "../lib", develop = true}
 
 [build-system]
 requires = ["poetry-core"]
