+++ language/pyproject.toml
@@ -9,7 +9,6 @@ package-mode = false
 [tool.poetry.dependencies]
 python = "^3.11"
 python-dotenv = "^1.0.0"
-httpx = "^0.24.1"
 aiofiles = "^23.2.1"
 azure-cognitiveservices-speech = "^1.34.1"
 pydub = "^0.25.1"
@@ -19,12 +18,14 @@ pytz = "^2023.3.post1"
 cachetools = "^5.3.2"
 pydantic = "^2.5.3"
 pydantic-settings = "^2.1.0"
-aiohttp = "^3.9.1"
 sqlalchemy = "^2.0.25"
 azure-storage-blob = "^12.19.0"
 azure-identity = "^1.15.0"
 tenacity = "^8.2.3"
 psycopg2-binary = "^2.9.9"
+
+[tool.poetry.group.dev.dependencies]
+lib = {path = "../lib", develop = true}
 
 [build-system]
 requires = ["poetry-core"]
