+++ lib/README.md
@@ -1,0 +1,1 @@
+# Common utility libraries for JBManager
