+++ lib/lib/azure_storage.py
