+++ lib/lib/file_storage/registry.py
