+++ lib/pyproject.toml
@@ -2,9 +2,8 @@
 name = "lib"
 version = "0.1.0"
 description = ""
-authors = ["Your Name <you@example.com>"]
+authors = ["Shrey Pandey <shreypandey1509@gmail.com>"]
 readme = "README.md"
-package-mode = false
 
 [tool.poetry.dependencies]
 python = "^3.11"
